[test] moving test class to test_file.py
